+++ packages/display3/src/index.ts
@@ -1,4 +1,5 @@
 export * from './Chronometer';
+export * from './controls/PanControl2D';
 export * from './Display';
 export * from './Stylesheets';
 export {getContentAreaSize} from './styleUtils';