+++ packages/tiled-maps/src/index.ts
@@ -1,5 +1,4 @@
 export * from './AABB2';
-export * from './controls/PanControl2D';
 export * from './IMap2DLayer';
 export * from './IMap2DTileDataProvider';
 export * from './IMap2DTileRenderer';